Create MyBooks.py
This is the code for an API written in Python using FLASK which allows the user to read the list of books on his/ her book shelf. It allows the user to add a new book to the list, update the rating of a book already shelved or delete a book from the shelf.